Rename cas3v2 references in symbols
