+++ e2e/tests/stepDefinitions/manage/bedspaceSearch.ts
@@ -5,8 +5,8 @@ import DashboardPage from '../../../../cypress_shared/pages/temporary-accommodat
 import BedspaceSearchPage from '../../../../cypress_shared/pages/temporary-accommodation/manage/bedspaceSearch'
 import {
   bedspaceSearchFormParametersFactory,
-  cas3v2BedspaceSearchResultFactory,
-  cas3v2BedspaceSearchResultsFactory,
+  cas3BedspaceSearchResultFactory,
+  cas3BedspaceSearchResultsFactory,
 } from '../../../../server/testutils/factories'
 import { characteristicsToSearchAttributes } from '../../../../cypress_shared/utils/bedspaceSearch'
 
@@ -55,9 +55,9 @@ Given('I search for a bedspace', () => {
 
 Then('I should see the bedspace search results', () => {
   cy.then(function _() {
-    const results = cas3v2BedspaceSearchResultsFactory.build({
+    const results = cas3BedspaceSearchResultsFactory.build({
       results: [
-        cas3v2BedspaceSearchResultFactory.forBedspace(this.premises, this.bedspace).build({
+        cas3BedspaceSearchResultFactory.forBedspace(this.premises, this.bedspace).build({
           overlaps: [],
         }),
       ],
